fixed performance regression due to state
This commit makes a couple of fixes to improve the computational
efficiency of the contracts. There was a performance regression when the
state was introduced (notably, `ExitStack` caused a lot of overhead).

This commit also introduces benchmarking into the continuous integration
(as part of the precommit.py script) so that we can continously monitor
the efficiency in the future.