In-lined assert and capture functions
We had to in-line `_assert_*` and `_capture` functions to improve
readability as separated functions lead to bugs which were hard to
reason about (as the logic of the call was separated in the source code
from its implementation).

This patch also prepares the ground for adding support for async
functions and async conditions.