Added support for error-as-instance
This patch allows the user to supply error as an instance of
``BaseException``. This is practical for cases where the error is fixed
and need not be re-instantiated on every violation.

An additional benefit is that the syntax is a bit shorter than with a
argument-less callable (one `lambda: ` less).