Fixed handling of `self` when passed as kwarg
This patch fixes a bug caused when `self` is passed as a keyword
arguments instead as a positional argument..

Usually, `self` is passed as a positional argument so this bug went
unnoticed for a long time. Notably, preconditions and postconditions
worked fine, but the problem surfaced when unbounding an instance method
and passing `self` as keyword argument explicitly.