Added reporting of all arguments on violation
Many times in production or debugging it was hard to discern what caused
the violation of the contract since only the condition arguments were
represented in the violation message. This is often not enough to
identify the test case or to give us a hint what might have provoked the
bug.

This patch represents all the arguments of the function call at the time
of violation so that the bugs are easier to trace.