Upgraded to mypy 0.910 and pylint 2.9.6 (#225)
This patch upgrades the mypy to 0.910. This was necessary so that we can
test the type inference with the latest mypy.

We also had to upgrade pylint 2.9.6 since the checks broke on CI due to
astroid having hard time parsing `_metaclass.py`. Due to this upgrade,
we had to disable all `no-member` warnings from pylint as pylint did not
properly infer types in condition functions.

Related to #224.