Fixed representation of numpy arrays
The numpy arrays do not like `PLACEHOLDER in arr` which will cause a
`ValueError`. However, we tested for `PLACEHOLDER` everywhere using
`in` operator in the representation module.

This patch fixes the problem by using explicit `any(...)` and checking
also explicitly for `... is PLACEHOLDER`. In contrast, `in` operator
uses equality (`==`) instead of identity (`is`), which was not the
desirable behavior anyhow originally.

Fixes #229.