Fixed mypy error on missing ``asttokens.ASTTokens``
Mypy complained that ``asttokens.ASTTokens`` is missing with the newest
asttokens version (2.0.7). Honestly, we are at loss what caused this
error as ``asttokens/__init__.py`` did not change. We tried updating
mypy to the latest version (0.971) which did not fix the error either.

We are leaving mypy set to the latest version and fix the mypy error by
explicitly writing ``asttokens.asttokens.ASTTokens`` instead of
``asttokens.ASTTokens``. While it is ugly, it is a quick fix for now.