Fixed in-progress set for async
Previously, we used `threading.local()` to keep track of contracts which
are in progress of verification to allow for recursive contracts. While
this works for threads, this approach does not work with async
coroutines.

In this fix we use `contextvars` instead of `threading.local()` to
manage the context, and keep track of concurrent contracts and function
calls.