Replaced YAPF with black
YAPF was giving us headaches as the reformattings were neither intuitive
nor predictable. We wanted to upgrade for Python 3.11, and tried to
upgrade YAPF to the latest version. As it caused a lot of reformatting,
we decided to drop YAPF and replace it with black.

Black seems a bit more stable, and subjectively gives us nicer
formatting.