Updated typeguard and deal to latest versions
We have to update the versions so that distributions can package them
with the latest dependencies. Originally, we thought that we can simply
separate the benchmark dependencies in a separate group in `setup.py`,
but it turned out that some of the tests actually did involve `deal` and
`typeguard`.

Additionally, we also fix the tests for the latest version of typeguard.

Fixes #282.