Tested for data classes with fields
In this patch, we test for `dataclasses` which use `field`'s to define
default values. Originally, we thought that there are issues with data
classes in general due to the issue #288, but it turned out that the
order of decorators matters (``dataclass`` first).

We tested everything, and documented the order of the decorators
accordingly.

Fixes #288.