Fix child invariants checked in `super().__init__` (#301)
Previously, we determined the invariants based on the `self` passed to
the function. However, in case of `super().__init__`, the invariants
that need to be checked after the call are those belonging to to
the super class, not the current (child) class.

This change fixes the issue by passing in the class corresponding to
the invariants alongside function and parameters, so that we can
properly resolve which invariants need to be checked.

Fixes #300.